labmanager/requirements.txt
asgiref==3.3.4
pynacl==1.3.0
djangorestframework==3.12.4
bcrypt==3.2.0
beautifulsoup4==4.9.3
argon2-cffi==20.1.0
bs4==0.0.1
certifi==2020.12.5
celery==4.4.2
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
cryptography==3.4.2
Django==3.1.7
django-bootstrap-form==3.4
django-bootstrap-v5==1.0.4
django-braces==1.14.0
django-celery-results==2.0.1
django-ckeditor==6.1.0
celery-progress==0.1.1
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-filter==2.4.0
django-simple-history
importlib-metadata==3.10.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.1
pip-search==0.0.7
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.9.0
pytz==2021.1
PyYAML==5.4.1
pyvmomi==7.0.2
paramiko==2.7.2
requests==2.25.1
misaka==2.1.1
rich==10.2.2
redis==3.5.3
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26

